Fixed warnings.

+++ Path.cpp
@@ -3,8 +3,8 @@
 
 const char* Path::FindFileName(const char* path)
 {
-    int length = strlen(path);
+    size_t length = strlen(path);
     const char* p = path + length - 1;
-    for (int i = length - 1; i >= 0 && *p != '/' && *p != '\\'; i--, p--);
+    for (size_t i = length - 1; i >= 0 && *p != '/' && *p != '\\'; i--, p--);
     return p + 1;
 }
